requirements.txt
httpx==0.27.2
openai==1.51.0
python-dotenv==1.0.1
gradio==4.44.1
overrides==7.7.0
zhipuai==2.1.5.20230904
icecream==2.1.3
python-pptx==1.0.2
pymysql==1.1.1
django==4.2.20
APScheduler==3.10.4
python-docx==1.1.2
markdown==3.7
unstructured==0.11.8
langchain-core==0.3.56
langchain-community==0.3.2
langchain-oceanbase==0.2.0
modelscope==1.17.0
addict==2.4.0
datasets==2.18.0
oss2==2.19.0
torch==2.6.0
simplejson==3.19.3
sortedcontainers==2.4.0
transformers==4.45.2
edge-tts>=7.0.0
pyahocorasick==2.1.0
py2neo==2021.2.4
PyPDF2==3.0.1
SpeechRecognition==3.10.4
opencc-python-reimplemented==0.1.7
faiss-cpu==1.9.0
soundfile==0.12.1
whisper==1.1.10
openai-whisper==20240930